+++ ruststep/src/ast/de/record.rs
@@ -1,30 +1,8 @@
-use crate::{ast::*, error::*, parser::exchange::simple_record};
+use crate::{ast::*, error::*};
 use serde::{
     de::{self, IntoDeserializer},
     forward_to_deserialize_any,
 };
-use std::str::FromStr;
-
-/// A struct typed in EXPRESS schema, e.g. `A(1.0, 2.0)`
-///
-/// ```
-/// use ruststep::ast::{Record, Parameter};
-/// use std::str::FromStr;
-///
-/// let record = Record::from_str("A(1, 2)").unwrap();
-/// assert_eq!(
-///     record,
-///     Record {
-///         name: "A".to_string(),
-///         parameter: Box::new(vec![Parameter::Integer(1), Parameter::Integer(2)].into())
-///     }
-/// )
-/// ```
-#[derive(Debug, Clone, PartialEq)]
-pub struct Record {
-    pub name: String,
-    pub parameter: Box<Parameter>,
-}
 
 impl<'de, 'record> de::Deserializer<'de> for &'record Record {
     type Error = Error;
@@ -40,18 +18,6 @@ impl<'de, 'record> de::Deserializer<'de> for &'record Record {
         bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64 char str string
         bytes byte_buf option unit unit_struct newtype_struct seq tuple
         struct tuple_struct map enum identifier ignored_any
-    }
-}
-
-impl FromStr for Record {
-    type Err = Error;
-    fn from_str(input: &str) -> Result<Self> {
-        use nom::Finish;
-        let input = input.trim();
-        let (_input, record) = simple_record(input)
-            .finish()
-            .map_err(|err| TokenizeFailed::new(input, err))?;
-        Ok(record)
     }
 }