+++ ruststep/tests/list.rs
@@ -1,5 +1,5 @@
 use nom::Finish;
-use ruststep::{ast::*, parser::exchange, place_holder::*};
+use ruststep::{ast::*, parser::exchange, tables::*};
 use serde::Deserialize;
 
 espr_derive::inline_express!(